+++ lib/rubocop/cop/rails/blank.rb
@@ -42,18 +42,13 @@ module RuboCop
         MSG_NOT_PRESENT = 'Use `%s` instead of `%s`.'.freeze
         MSG_UNLESS_PRESENT = 'Use `if %s` instead of `%s`.'.freeze
 
-        # `(send nil $_)` is not actually a valid match for an offense. Nodes
-        # that have a single method call on the left hand side
-        # (`bar || foo.empty?`) will blow up when checking
-        # `(send (:nil) :== $_)`.
         def_node_matcher :nil_or_empty?, <<-PATTERN
           (or
               {
                 (send $_ :!)
                 (send $_ :nil?)
-                (send $_ :== (:nil))
-                (send nil $_)
-                (send (:nil) :== $_)
+                (send $_ :== nil)
+                (send nil :== $_)
               }
               {
                 (send $_ :empty?)
@@ -65,7 +60,7 @@ module RuboCop
         def_node_matcher :not_present?, '(send (send $_ :present?) :!)'
 
         def_node_matcher :unless_present?, <<-PATTERN
-          (:if $(send $_ :present?) {nil (...)} ...)
+          (:if $(send $_ :present?) {nil? (...)} ...)
         PATTERN
 
         def on_send(node)